// src/types/partials/ParametersParameter.ts
export type TParametersParameter = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    name: String;
    valueBase64Binary?: TBase64Binary;
    valueBoolean?: Boolean;
    valueCanonical?: TCanonical;
    valueCode?: String;
    valueDate?: TDate;
    valueDateTime?: TDateTime;
    valueDecimal?: TDecimal;
    valueId?: TId;
    valueInstant?: TInstant;
    valueInteger?: TInt;
    valueMarkdown?: TMarkdown;
    valueOid?: TOid;
    valuePositiveInt?: TInt;
    valueString?: String;
    valueTime?: TTime;
    valueUnsignedInt?: TUnsignedInt;
    valueUri?: TUri;
    valueUrl?: TUrl;
    valueUuid?: TUuid;
    valueAddress?: TAddress;
    valueAge?: TQuantity;
    valueAnnotation?: TAnnotation;
    valueAttachment?: TAttachment;
    valueCodeableConcept?: TCodeableConcept;
    valueCoding?: TCoding;
    valueContactPoint?: TContactPoint;
    valueCount?: TQuantity;
    valueDistance?: TQuantity;
    valueDuration?: TQuantity;
    valueHumanName?: THumanName;
    valueIdentifier?: TIdentifier;
    valueMoney?: TMoney;
    valuePeriod?: TPeriod;
    valueQuantity?: TQuantity;
    valueRange?: TRange;
    valueRatio?: TRatio;
    valueReference?: TReference;
    valueSampledData?: TSampledData;
    valueSignature?: TSignature;
    valueTiming?: TTiming;
    valueContactDetail?: TContactDetail;
    valueContributor?: TContributor;
    valueDataRequirement?: TDataRequirement;
    valueExpression?: TExpression;
    valueParameterDefinition?: TParameterDefinition;
    valueRelatedArtifact?: TRelatedArtifact;
    valueTriggerDefinition?: TTriggerDefinition;
    valueUsageContext?: TUsageContext;
    valueDosage?: TDosage;
    valueMeta?: TMeta;
    resource?: TResourceContainer;
    part?: TParametersParameter[];
};